Moving comand line options to unity transport
