+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkManager.cs
@@ -976,18 +976,18 @@ namespace Unity.Netcode
         internal NetworkConnectionManager ConnectionManager = new NetworkConnectionManager();
         internal NetworkMessageManager MessageManager = null;
 
-        internal struct Override<T>
-        {
-            private T m_Value;
-            public bool Overidden { get; private set; }
-            internal T Value
-            {
-                get { return Overidden ? m_Value : default(T); }
-                set { Overidden = true; m_Value = value; }
-            }
-        };
-
-        internal Override<ushort> PortOverride;
+        // internal struct Override<T>
+        // {
+        //     private T m_Value;
+        //     public bool Overidden { get; private set; }
+        //     internal T Value
+        //     {
+        //         get { return Overidden ? m_Value : default(T); }
+        //         set { Overidden = true; m_Value = value; }
+        //     }
+        // };
+        //
+        // internal Override<ushort> PortOverride;
 
         /// <summary>
         /// Determines if the NetworkManager's GameObject is parented under another GameObject and
@@ -1168,7 +1168,7 @@ namespace Unity.Netcode
                 return;
             }
 
-            ParseCommandLineOptions();
+            //ParseCommandLineOptions();
 
             if (NetworkConfig.NetworkTransport == null)
             {
@@ -1740,39 +1740,39 @@ namespace Unity.Netcode
 #endif
         }
 
-        // Command line options
-        private const string k_OverridePortArg = "-port";
-
-        private string GetArg(string[] commandLineArgs, string arg)
-        {
-            var argIndex = Array.IndexOf(commandLineArgs, arg);
-            if (argIndex >= 0 && argIndex < commandLineArgs.Length - 1)
-            {
-                return commandLineArgs[argIndex + 1];
-            }
-
-            return null;
-        }
-
-        private void ParseArg<T>(string arg, ref Override<T> value)
-        {
-            if (GetArg(Environment.GetCommandLineArgs(), arg) is string argValue)
-            {
-                value.Value = (T)Convert.ChangeType(argValue, typeof(T));
-            }
-        }
-
-        private void ParseCommandLineOptions()
-        {
-#if UNITY_SERVER && UNITY_DEDICATED_SERVER_ARGUMENTS_PRESENT
-            if ( UnityEngine.DedicatedServer.Arguments.Port != null)
-            {
-                PortOverride.Value = (ushort)UnityEngine.DedicatedServer.Arguments.Port;
-            }
-#else
-            ParseArg(k_OverridePortArg, ref PortOverride);
-#endif
-        }
+//         // Command line options
+//         private const string k_OverridePortArg = "-port";
+//
+//         private string GetArg(string[] commandLineArgs, string arg)
+//         {
+//             var argIndex = Array.IndexOf(commandLineArgs, arg);
+//             if (argIndex >= 0 && argIndex < commandLineArgs.Length - 1)
+//             {
+//                 return commandLineArgs[argIndex + 1];
+//             }
+//
+//             return null;
+//         }
+//
+//         private void ParseArg<T>(string arg, ref Override<T> value)
+//         {
+//             if (GetArg(Environment.GetCommandLineArgs(), arg) is string argValue)
+//             {
+//                 value.Value = (T)Convert.ChangeType(argValue, typeof(T));
+//             }
+//         }
+//
+//         private void ParseCommandLineOptions()
+//         {
+// #if UNITY_SERVER && UNITY_DEDICATED_SERVER_ARGUMENTS_PRESENT
+//             if ( UnityEngine.DedicatedServer.Arguments.Port != null)
+//             {
+//                 PortOverride.Value = (ushort)UnityEngine.DedicatedServer.Arguments.Port;
+//             }
+// #else
+//             ParseArg(k_OverridePortArg, ref PortOverride);
+// #endif
+//         }
 
 #if UNITY_EDITOR
         internal static INetworkManagerHelper NetworkManagerHelper;
