+++ testproject/Assets/Scenes/CommandLineOverrideTest.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9a4c20895a6e8e4458ac97be10fc7ffd
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
