+++ testproject/Assets/Scripts/ServerTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9c6424f91fc70544190684bfc288041c
